requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
gunicorn